Audit all usages of Tracer.Instance and make sure all publicly accessible code paths use TracerProvider.GetTracer() instead. This includes several API's:
CorrelationIdentifier
This API requests the static value of the service/env/version/trace_id/span_id, so its calls to Tracer.Instance were replaced with TracerProvider.GetTracerInstance (which now returns an IInternalTracer API) and the IInternalTracer API implements the public ITracer API, in addition to providing the tracer's service/environment/version (its implementation accesses its settings object).

SpanContextExtractor / SpanContextInjector
This one was tricky. First, SpanContextInjector did not have a matching interface, so I created ISpanContextInjector. Then, the constructors for each implementation type were replaced with a static factory method to return either the type from the manual assembly or the automatic assembly. Then, there were various DuckType calls and new objects allocated to ensure that the automatic tracer could receiver or return an ISpanContext that the app knew about. And lastly the ISpanContext interface was expanded to include a SamplingPriority property, which is accepted in the public constructor of SpanContext.