Update ITracer and IInternalTracer interfaces to allow us to pass in an ISpanContext from the manual side to the automatic side for the StartActive overload with a SpanCreationSettings object.
From the manual side, the SpanContext will be passed in as-is if it originated from the manual assembly or unwrapped from the IDuckType that originated from the automatic assembly.

From the automatic side, the SpanContext will be used as-is if it originated from the automatic assembly or cloned into a local SpanContext if the argument implements the ISpanContext duck type.