[Incomplete] Clean up the public API for ITest.
This is  difficult because we currently have structs and enums in our public API. We might consider removing the structs so we don't have to decompose them on the way into methods where they are used. For the TestStatus enum, I have casted this into an int and back to pass it as a parameter.

See the in-line comment at tracer/src/Datadog.Trace/Ci/Test.cs : Line 192

One potential solution is to change the structs in the public interface into a class so we can duck type it.